+++ src/main/java/chess/domain/game/ChessGame.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
