+++ src/main/java/chess/domain/movingStrategy/KnightMovingStrategy.java
@@ -1,8 +1,8 @@
 package chess.domain.movingStrategy;
 
-import chess.domain.File;
-import chess.domain.Position;
-import chess.domain.Rank;
+import chess.domain.game.File;
+import chess.domain.game.Position;
+import chess.domain.game.Rank;
 
 public abstract class KnightMovingStrategy implements MovingStrategy {
