+++ src/main/java/chess/domain/movingStrategy/MoveDownDown.java
@@ -1,7 +1,7 @@
 package chess.domain.movingStrategy;
 
-import chess.domain.Position;
-import chess.domain.Rank;
+import chess.domain.game.Position;
+import chess.domain.game.Rank;
 
 public class MoveDownDown implements MovingStrategy {
