+++ src/main/java/chess/domain/movingStrategy/MovingStrategies.java
@@ -1,6 +1,6 @@
 package chess.domain.movingStrategy;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 
 import java.util.List;
