+++ src/main/java/chess/domain/piece/pawn/Pawn.java
@@ -1,6 +1,6 @@
 package chess.domain.piece.pawn;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 import chess.domain.movingStrategy.MovingStrategies;
 import chess.domain.movingStrategy.MovingStrategy;
 import chess.domain.piece.Color;
